requirements.txt
redis==4.3.4
python-dotenv~=0.21.0
vk-api==11.9.9
python-telegram-bot==13.14


